Added Create Badges Route and moved Data components in it.
