+++ frontend/app/templates/components/welcome-header.hbs
@@ -1,4 +1,6 @@
 <div class="ui content">
   <h1 class="weight-300 lead">The simplest way to create Badges.</h1>
-  <a class="ui orange button" style="text-transform: uppercase; font-weight: 400">Create Badges</a>
+  {{#link-to 'create-badges'}}
+    <a class="ui orange button" style="text-transform: uppercase; font-weight: 400">Create Badges</a>
+  {{/link-to}}
 </div>
